--- src/helpers/data/index.ts
@@ -1,6 +1,0 @@
-export * from './calculate-distances';
-export * from './filter-meeting-data';
-export * from './flatten-and-sort-indexes';
-export * from './get-index-by-key';
-export * from './load-meeting-data';
-export * from './translate-google-sheet';
