+++ src/helpers/format-array.ts
