+++ src/helpers/format-string.ts
