--- src/helpers/format/index.ts
@@ -1,10 +1,0 @@
-export * from './format-address';
-export * from './format-array';
-export * from './format-classes';
-export * from './format-conference-provider';
-export * from './format-directions-url';
-export * from './format-feedback-email';
-export * from './format-ics';
-export * from './format-slug';
-export * from './format-string';
-export * from './format-url';
