+++ src/helpers/translate-google-sheet.ts
@@ -1,9 +1,9 @@
 import { DateTime } from 'luxon';
 
-import type { JSONData } from '../../types';
-import { formatSlug } from '../format';
-import { en, es, fr, ja, sv } from '../../i18n';
-import { settings } from '../../helpers';
+import type { JSONData } from '../types';
+import { formatSlug } from './format-slug';
+import { en, es, fr, ja, sv } from '../i18n';
+import { settings } from './settings';
 
 export type GoogleSheetData = {
   values: string[][];
